# Optimize string functions with shrn instruction