We found that string functions were using AND+ADDP
to find the nibble/syndrome mask but there is an easier
opportunity through `SHRN dst, src, 4` and has same
latency on all SIMD ARMv8 targets as ADDP. There are also
gaps for memcmp but that's probably for another patch

We see 10-20% savings for small-mid size cases which are
primary cases for general workloads 

Repo benchmarks for strlen on Neoverse-N1:

New:

```
Random strlen (bytes/ns):
                strlen 13180875.00
      __strlen_aarch64 1.77
  __strlen_aarch64_mte 1.40

Small aligned strlen (bytes/ns):
                strlen 1B: 0.33 2B: 0.67 4B: 1.20 8B: 2.66 16B: 5.33 32B: 6.84 64B: 10.09 
      __strlen_aarch64 1B: 0.45 2B: 0.90 4B: 1.71 8B: 3.59 16B: 7.18 32B: 10.11 64B: 14.75 
  __strlen_aarch64_mte 1B: 0.60 2B: 1.20 4B: 2.40 8B: 4.79 16B: 9.59 32B: 13.69 64B: 17.43 

Small unaligned strlen (bytes/ns):
                strlen 1B: 0.28 2B: 0.55 4B: 1.09 8B: 2.21 16B: 4.36 32B: 6.28 64B: 10.08 
      __strlen_aarch64 1B: 0.39 2B: 0.79 4B: 1.58 8B: 3.16 16B: 6.31 32B: 9.01 64B: 12.77 
  __strlen_aarch64_mte 1B: 0.60 2B: 1.20 4B: 2.40 8B: 3.44 16B: 6.69 32B: 8.70 64B: 15.95 

Medium strlen (bytes/ns):
                strlen 128B: 14.73 256B: 20.15 512B: 23.03 1K: 24.87 2K: 25.88 4K: 25.45 
      __strlen_aarch64 128B: 23.53 256B: 24.59 512B: 32.28 1K: 36.96 2K: 39.59 4K: 38.19 
  __strlen_aarch64_mte 128B: 23.91 256B: 26.54 512B: 28.85 1K: 28.94 2K: 29.56 4K: 29.83 
```

Old:


```
Random strlen (bytes/ns):
                strlen 26361750.00
      __strlen_aarch64 1.77
  __strlen_aarch64_mte 1.25

Small aligned strlen (bytes/ns):
                strlen 1B: 0.33 2B: 0.67 4B: 1.20 8B: 2.66 16B: 5.33 32B: 6.85 64B: 10.09 
      __strlen_aarch64 1B: 0.45 2B: 0.90 4B: 1.71 8B: 3.59 16B: 7.18 32B: 10.10 64B: 14.74 
  __strlen_aarch64_mte 1B: 0.60 2B: 1.20 4B: 2.18 8B: 4.79 16B: 9.58 32B: 11.95 64B: 17.37 

Small unaligned strlen (bytes/ns):
                strlen 1B: 0.28 2B: 0.55 4B: 1.09 8B: 2.20 16B: 4.35 32B: 6.28 64B: 10.09 
      __strlen_aarch64 1B: 0.40 2B: 0.79 4B: 1.58 8B: 3.16 16B: 6.31 32B: 9.00 64B: 12.78 
  __strlen_aarch64_mte 1B: 0.60 2B: 1.20 4B: 2.18 8B: 3.13 16B: 5.99 32B: 8.71 64B: 15.48 

Medium strlen (bytes/ns):
                strlen 128B: 14.74 256B: 20.15 512B: 23.07 1K: 24.93 2K: 25.89 4K: 25.99 
      __strlen_aarch64 128B: 23.56 256B: 25.18 512B: 33.03 1K: 37.08 2K: 39.61 4K: 38.27 
  __strlen_aarch64_mte 128B: 22.23 256B: 26.05 512B: 27.97 1K: 29.00 2K: 29.64 4K: 29.69
```