@danlark1 2022-06-22
I have also published it to glibc https://patchwork.sourceware.org/project/glibc/patch/20220622072946.3123916-1-danilak@google.com/